+++ src/uiprotect/data/types.py
@@ -436,6 +436,7 @@ class VideoMode(str, ValuesEnumMixin, enum.Enum):
     SPORT = "sport"
     SLOW_SHUTTER = "slowShutter"
     LPR_NONE_REFLEX = "lprNoneReflex"
+    LPR_REFLEX = "lprReflex"
     # should only be for unadopted devices
     UNKNOWN = "unknown"
 
@@ -448,6 +449,7 @@ class AudioStyle(str, UnknownValuesEnumMixin, enum.Enum):
 
 @enum.unique
 class RecordingMode(str, ValuesEnumMixin, enum.Enum):
+    ADAPTIVE = "adaptive"
     ALWAYS = "always"
     NEVER = "never"
     SCHEDULE = "schedule"